refactor(desktop): decompose tool summary utilities and add file-edit tests
Split tool-summary-builder responsibilities into focused summary, duration, file-edit, and shared helper modules while preserving existing exports. Add dedicated file-edit utility test coverage in agent-chat-message-card-model.test.ts.
